Move log level config to trait for reusing
